Move SignedBigNum to cfx_type and Implement Tree-Array
