+++ core/Cargo.toml
@@ -38,6 +38,7 @@ lru = "0.1.11"
 threadpool = "1.0"
 rayon = "1.0"
 link-cut-tree = { path = "../util/link-cut-tree" }
+tree-array = { path = "../util/tree-array" }
 priority-send-queue = { path = "../util/priority-send-queue" }
 toml = "0.4"
 metrics = { path = "../util/metrics" }
